src/llamaindex/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
llama-index>=0.11.0
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-embeddings-text-embeddings-inference
llama-index-vector-stores-qdrant
llama-index-storage-docstore-redis
qdrant-client>=1.12.0
fastembed>=0.4.0
minio>=7.2.0
python-keycloak>=3.0.0
python-multipart==0.0.9
pydantic>=2.6.1
pydantic-settings>=2.0.0
requests==2.31.0
